Divided modules and added lazy loading
