+++ backend/routes/posts.js
@@ -1,34 +1,16 @@
 const express = require("express");
-const multer = require('multer');
+
+
 const PostsController = require('../controllers/posts');
 
 const checkAuth = require('../middleware/check-auth');
+const extractFile = require('../middleware/file');
 
 const router = express.Router();
-const MIME_TYPE_MAP = {
-  'img/png': 'png',
-  'img/jpeg': 'jpg',
-  'img/jpg': 'jpg',
-}
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    const isValid = MIME_TYPE_MAP[file.mimetype];
-    let error = new Error('Invalid mime type');
-    if(isValid) {
-      error= null;
-    }
-    cb(null, 'backend/images');
-  },
-  filename: (req, file, cb) => {
-    const name = file.originalname.toLocaleLowerCase().split(' ').join('-');
-    const ext = MIME_TYPE_MAP[file.mimetype];
-    cb(null, name + '-' + Date.now() + '.' + ext);
-  }
-});
 
-router.post('', checkAuth, multer({storage: storage}).single('image'), PostsController.createPost);
+router.post('', checkAuth, extractFile, PostsController.createPost);
 
-router.put('/:id', checkAuth, multer({storage: storage}).single('image'), PostsController.updatePost);
+router.put('/:id', checkAuth, extractFile, PostsController.updatePost);
 
 router.get('', PostsController.getPosts);
