+++ src/app/angular-material.module.ts
@@ -1,0 +1,25 @@
+import { NgModule } from '@angular/core';
+import {
+  MatInputModule,
+  MatCardModule,
+  MatButtonModule,
+  MatToolbarModule,
+  MatExpansionModule,
+  MatProgressSpinnerModule,
+  MatPaginatorModule,
+  MatDialogModule
+} from '@angular/material';
+
+@NgModule({ // Normally we need to import the dependencies first, but omitting things we export auto imports them by NG
+  exports: [
+    MatInputModule,
+    MatCardModule,
+    MatButtonModule,
+    MatToolbarModule,
+    MatExpansionModule,
+    MatProgressSpinnerModule,
+    MatPaginatorModule,
+    MatDialogModule
+  ]
+})
+export class AngularMaterialModule {}
