+++ src/app/app-routing.module.ts
@@ -3,16 +3,13 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { PostListComponent } from './posts/post-list/post-list.component';
 import { PostCreateComponent } from './posts/post-create/post-create.component';
-import { LoginComponent } from './auth/login/login.component';
-import { SignUpComponent } from './auth/signup/signup.component';
 import { AuthGuard } from './auth/auth.guard';
 
 const routes: Routes = [
   { path: '', component: PostListComponent },
   { path: 'create', component: PostCreateComponent, canActivate: [AuthGuard] },
   { path: 'edit/:postId', component: PostCreateComponent, canActivate: [AuthGuard] },
-  { path: 'login', component: LoginComponent},
-  { path: 'signup', component: SignUpComponent}
+  { path: 'auth', loadChildren: './auth/auth.module#AuthModule' } // lazy loading modules
 ];
 
 @NgModule({
