+++ src/app/app.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Post } from './posts/post.model';
 import { AuthService } from './auth/auth.service';
 
 @Component({
