+++ src/app/auth/auth-routing.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { LoginComponent } from './login/login.component';
+import { SignUpComponent } from './signup/signup.component';
+
+const routes: Routes = [
+  { path: 'login', component: LoginComponent},
+  { path: 'signup', component: SignUpComponent}
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(routes)
+  ],
+  exports: [RouterModule]
+})
+export class AuthRoutingModule {}
