+++ src/app/auth/auth.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+import { LoginComponent } from './login/login.component';
+import { SignUpComponent } from './signup/signup.component';
+import { AngularMaterialModule } from '../angular-material.module';
+import { AuthRoutingModule } from './auth-routing.module';
+
+@NgModule({
+  declarations: [
+    LoginComponent,
+    SignUpComponent
+  ],
+  imports: [
+    CommonModule,
+    AngularMaterialModule,
+    FormsModule,
+    AuthRoutingModule
+  ]
+})
+export class AuthModule {}
