+++ src/app/header/header.component.html
@@ -8,10 +8,10 @@
       <a mat-button [routerLink]="['/create']" routerLinkActive="mat-accent"> New Post </a>
     </li>
     <li *ngIf="!userIsAuthenticated">
-      <a mat-button [routerLink]="['/login']" routerLinkActive="mat-accent"> Login </a>
+      <a mat-button [routerLink]="['/auth/login']" routerLinkActive="mat-accent"> Login </a>
     </li>
     <li *ngIf="!userIsAuthenticated">
-      <a mat-button [routerLink]="['/signup']" routerLinkActive="mat-accent"> Signup </a>
+      <a mat-button [routerLink]="['/auth/signup']" routerLinkActive="mat-accent"> Signup </a>
     </li>
     <li *ngIf="userIsAuthenticated">
       <button mat-button (click)="OnLogOut()">Logout</button>
