+++ src/app/posts/posts.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+import { AngularMaterialModule } from '../angular-material.module';
+import { PostCreateComponent } from './post-create/post-create.component';
+import { PostListComponent } from './post-list/post-list.component';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+
+
+@NgModule({
+  declarations: [
+    PostCreateComponent,
+    PostListComponent
+  ],
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    AngularMaterialModule,
+    RouterModule
+  ]
+})
+export class PostsModule {}
